+++ core/list_formula_attachments_data.php
@@ -20,6 +20,7 @@ foreach ($docs as $doc) {
 	$r['type'] = (int)$doc['type'];
 	$r['name'] = (string)$doc['name']?:'N/A';
 	$r['notes'] = (string)$doc['notes']?:'N/A';
+	$r['created'] = (string)$doc['created']?:'N/A';
 	$r['docData'] = (string)$doc['docData'];
 	$r['docSize'] = (string)formatBytes(strlen($doc['docData']));