+++ releasenotes.md
@@ -1,3 +1,4 @@
 Whats New in v10.0
 --------------------------
+- Minor improvements in formula attachments section
 - For full details please refer to the CHANGELOG